Working REST API version with FASTAPI, postgreSQL database

+++ app/routers/post.py
@@ -1,4 +1,3 @@
-from datetime import datetime
 import uuid
 from .. import schemas, models
 from sqlalchemy.orm import Session
@@ -54,23 +53,20 @@ def update_post(
     user_id: str = Depends(require_user),
 ):
     post_query = db.query(models.Post).filter(models.Post.id == id)
-    db_post = post_query.first()
+    updated_post = post_query.first()
 
-    if not db_post:
+    if not updated_post:
         raise HTTPException(
             status_code=status.HTTP_200_OK, detail=f"No post with this id: {id} found"
         )
-    if db_post.user_id != uuid.UUID(user_id):
+    if updated_post.user_id != uuid.UUID(user_id):
         raise HTTPException(
             status_code=status.HTTP_403_FORBIDDEN,
             detail="You are not allowed to perform this action",
         )
-    post.user_id = db_post.user_id
-    post.created_at = db_post.created_at
-    post.updated_at = datetime.utcnow()
-    post_query.update(post.dict(exclude_none=True), synchronize_session=False)
+    post_query.update(post.dict(), synchronize_session=False)
     db.commit()
-    return db_post
+    return updated_post
 
 
 @router.get("/{id}", response_model=schemas.PostResponse)
@@ -96,6 +92,12 @@ def delete_post(
         raise HTTPException(
             status_code=status.HTTP_404_NOT_FOUND,
             detail=f"No post with this id: {id} found",
+        )
+
+    if post.owner_id != user_id:
+        raise HTTPException(
+            status_code=status.HTTP_403_FORBIDDEN,
+            detail="You are not allowed to perform this action",
         )
     post_query.delete(synchronize_session=False)
     db.commit()
